test: differentiate npm_audit unit from integration tests
